Day 16: Abort search if it's physically impossible to succeed
It's quite easy to check the maximum possible potential by counting up
the pressure released if all remaining valves are opened at this
point. While this is impossible most of the time, if that doesn't
surpass the current highest pressure released, we don't need to bother
enumerating the different options.