More aggressively reduce day 16 search space
This is slowly approaching reasonable run time.

First, factor in the time it would take to get to a given valve to
calculate the potential pressure released on the current path. This
cuts out paths that have high-flow valves closed far away from the
agents.

Also delve into higher-flow paths earlier by sorting the options that
way. This increases the chances that we can exclude large parts of the
search space early on by getting to at least 80-90% of the maximum
pressure released.